+++ react/src/index.js
@@ -78,6 +78,7 @@ window.FolioConsole.React.init = (domRoot) => {
       url: domRoot.dataset.url,
       options: domRoot.dataset.options ? JSON.parse(domRoot.dataset.options) : null,
       sortable: domRoot.dataset.sortable !== '0',
+      maxItems: domRoot.dataset.maxItems ? parseInt(domRoot.dataset.maxItems) : null,
       atomSetting: domRoot.dataset.atomSetting,
       menuPlacement: domRoot.dataset.menuPlacement,
       serialization: domRoot.dataset.serialization,